Major Level Changes and Bug Fixes
The following changes are the result of Monday August 24th’s play test….
-Updated the tutorial level with text to teach the player and minor
adjustments to the flow of the level.
-Added checkpoints and moved the engine in level 02 so the player can
actually see it.
-Also moved the wall run platforms to better indicate where to run on
level 02
-Removed and fixed the bug where players could land and “float” on top
of an invisible platform in level 03
-Also added a wall between the large spinning platforms to stop players
from cheating
-Updated level four to with better checkpoint and anti-cheating walls
to prevent the player from skipping to the finish.
-Also on level four, the ground now successfully kills the player
keeping them from being stranded outside the level.
-Tutorial now has a message and blocking volume to prevent the player
from continuing in the unfinished portion of the level.