Section 1 & 2 of level 1 complete
Updated Pivot points of Pendulums.  Recreated prefabs to exist without
helper (empty) game objects.  Constructed first 2 sections of level 1,
and they are fully playable.  (In effect this can be easily connected to
the existing level 1 as an intro section).  However, I'm working on
taking Kyle's new layout for level 1 and re-building it for class
tonight.